Search: Add suggestions
This change adds suggestions to search. It revamps a number of aspects of search along with it:

* Suggestions for any text input (search bar or limit.
* Search does prefix searching now, not any substring.
* Search by enum renamed to SearchableField and NameOrCode added to it.
* Some rounded corner fixes to dropdown component styling
* "SearchHighlighted" component refactored into smaller pieces that can be reused in sections. Now called "ObjectField"
* Fixed highlight appearance to make sure the text is still legible.

I'll note that it does seem to hurt page performance. I tried to mitigate that by using an async method but it still seems to slow down a bit -- I'll have to investigate why.