Language data: Add glottolog codes for remaining language families
There are a number of language families that did not have glottocodes yet. I noticed this as I was adding data for from the Canadian Census. This updates the manual mapping of glottocodes to ISO codes to fill in the remaining languages.

There are also a few other updates in a few places to improve a few things that helped me identify the correct glottocodes.