Censuses: Add new object type for population records
One service we want to deliver is precise population estimates. To do
that I created a new object type for a Census. Now these collections may
be government Censuses or academic studies or something else -- but the
word Census really just means counting people, so I think it works and
its nice and short.

Apologies that this PR is so big -- it's a lot to review. You can do it
commit by commit. That's also why I constrained the new data to only
come from the Canadian census.

The biggest changes are:
* Adding a Canada census import for the 2021 census `ca2021.tsv`.
* Add a readme describing how add census data.
* Add a file `CensusData.tsv` for importing and parsing censuses.
* Add a new ObjectType `Census` and update downstream ObjectType and
ObjectData references.
* Adding a whole suite of `/view/census/*.tsx` files that visualize the
census information (see below).

|View|Screenshot|
|--|--|
|Details|![Screenshot 2025-05-30 at 18 58
55](https://github.com/user-attachments/assets/d636fe90-8299-46ca-bcfd-d77fcdcb0260)
|Card List|![Screenshot 2025-05-30 at 18 55
09](https://github.com/user-attachments/assets/2801e476-f146-458f-b6ac-66b36f976004)|
|Hierarchy|![Screenshot 2025-05-30 at 18 55
19](https://github.com/user-attachments/assets/8f95b253-5511-4e40-89c8-f5658ff33fff)
|Table|![Screenshot 2025-05-30 at 18 55
25](https://github.com/user-attachments/assets/ff0991a8-9935-4f6b-8194-f139d9c6cb26)