Censuses: Add Nepali census (#52)
This change adds the 2021 Nepali census. Unfortunately, there are a
number if language families (eg. Tharu, Tamang) and a number of uncoded
dialects (mostly of Nepali but not always). Inferring individual
Tharu/Tamang language estimates or adding partial dialects into better
language estimates needs a bit more manual curation though so we will
have to plan for that separately.

While I did this I also added 2 more good fields to the census metadata:
collectorType and collectorName.


![Screenshot 2025-06-05 at 12 12
13](https://github.com/user-attachments/assets/e956daa6-c524-41bf-af14-2bbf2e071648)

See the 4 new tables for Nepal. The live link for this view is
https://translation-commons.github.io/lang-nav/?objectType=Census&view=Table
![Screenshot 2025-06-05 at 12 43
21](https://github.com/user-attachments/assets/8c7cdd15-dbe7-4d66-9be9-b23781dce954)