Population: Add estimate for how many people write each language (#53)
Most population estimates are for the number of people that speak a
language -- but some people that want this data want to know how many
people use their phones or keyboards in the language. While we do not
yet have a direct source for written literacy, this approximates it
using the national literacy rate.

Example from this URL

https://translation-commons.github.io/lang-nav/?view=Details&objectType=Locale&objectID=mai_NP

|Before|After|
|--|--|
|![Screenshot 2025-06-06 at 14 02
21](https://github.com/user-attachments/assets/273f7ede-2f1b-436e-a1c2-35aca37d693c)|![Screenshot
2025-06-06 at 14 02
27](https://github.com/user-attachments/assets/55b36217-e5d3-46a9-b6ef-a377a3b5b365)|